Initial commit
Includes following:

* ANTECKNINGAR.md where I have collected all useful ocaml knowledge and
broken it down so far. Will probably reorganize this a bit down the line
as this file grows in size.

* "dune_project" is a folder including a simple dune project example.
see information in ANTECKNINGAR.md on how to step-by-step setup this
kind of project

* "simple_back_script_build" is a folder with a simple bash script
"build.sh" that is written from scratch showing how to build ocaml
code by directly calling the ocaml compiler (ocamlopt), as suppose
to relying on the dune build-system.